refactor types and update logic for new styles Fixes #105
