+++ src/libs/type/index.ts
@@ -3,5 +3,12 @@ export { aspectList } from './lib/aspect'
 export { iconList } from './lib/icon'
 
 // types
-export type { TIconList } from './lib/icon'
 export type { TAspect } from './lib/aspect'
+export type { TButton } from './lib/button'
+export type { TFlexBox } from './lib/flexbox'
+export type { TGridBox } from './lib/gridbox'
+export type { TIcon } from './lib/icon'
+export type { TImage } from './lib/image'
+export type { TInput } from './lib/input'
+export type { TIconList } from './lib/icon'
+export type { TText } from './lib/text'
