+++ src/libs/type/lib/button.ts
@@ -1,0 +1,18 @@
+export type TButton = {
+  // Costum properites
+  disabled?: boolean;
+  padding?: number | number[];
+  fontSize?: number;
+  type?: 'button' | 'submit' | 'reset';
+  color?: string;
+  backgroundColor?: string;
+  width?: string;
+  height?: string;
+  border?: string;
+  cursor?: string;
+  borderRadius?: string;
+  gap?: number | number[];
+  boxShadow?: string;
+  variant?: string;
+  children?: React.ReactNode;
+};
