+++ src/libs/ui/atom/gridbox/gridbox.type.ts
@@ -1,15 +1,15 @@
-export type TGridBox = {
-  children: React.ReactNode
-  gridTemplateColumns?: string
-  gridTemplateRows?: string
-  gap?: number
-  justifyItems?: string
-  alignItems?: string
-  justifyContent?: string
-  alignContent?: string
-  gridAutoFlow?: string
-  gridAutoColumns?: string
-  gridAutoRows?: string
-  placeItems?: string
-  placeContent?: string
-}
+// export type TGridBox = {
+//   children: React.ReactNode
+//   gridTemplateColumns?: string
+//   gridTemplateRows?: string
+//   gap?: number
+//   justifyItems?: string
+//   alignItems?: string
+//   justifyContent?: string
+//   alignContent?: string
+//   gridAutoFlow?: string
+//   gridAutoColumns?: string
+//   gridAutoRows?: string
+//   placeItems?: string
+//   placeContent?: string
+// }
