+++ src/libs/ui/atom/icon/icon.type.ts
@@ -1,14 +1,14 @@
-import { TIconList } from '@/libs/type'
+// import { TIconList } from '@/libs/type'
 
-export type TIcon = {
-  icon?: TIconList
-  backgroundColor?: string
-  fontSize?: number
-  width?: string
-  height?: string
-  borderRadius?: string
-  round?: boolean
-  altText?: string
-  color?: string
-  boxShadow?: string
-}
+// export type TIcon = {
+//   icon?: TIconList
+//   backgroundColor?: string
+//   fontSize?: number
+//   width?: string
+//   height?: string
+//   borderRadius?: string
+//   round?: boolean
+//   altText?: string
+//   color?: string
+//   boxShadow?: string
+// }
