+++ src/libs/ui/atom/input/input.type.ts
@@ -1,26 +1,26 @@
-import { FormikProps } from 'formik';
+// import { FormikProps } from 'formik';
 
-export type TInput = {
-  children?: React.ReactNode;
-  label?: string;
-  describedBy?: string;
-  color?: string;
-  backgroundColor?: string;
-  fontSize?: number;
-  padding?: number | number[];
-  width?: string;
-  height?: string;
-  border?: string;
-  borderRadius?: string;
-  field?: {
-    name: string;
-    value: string;
-    onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-    onBlur: (e: React.FocusEvent<HTMLInputElement>) => void;
-  };
-  form?: FormikProps<Record<string, any>>;
-  id?: string;
-  type?: string;
-  placeholder?: string;
-  required?: boolean;
-};
+// export type TInput = {
+//   children?: React.ReactNode;
+//   label?: string;
+//   describedBy?: string;
+//   color?: string;
+//   backgroundColor?: string;
+//   fontSize?: number;
+//   padding?: number | number[];
+//   width?: string;
+//   height?: string;
+//   border?: string;
+//   borderRadius?: string;
+//   field?: {
+//     name: string;
+//     value: string;
+//     onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+//     onBlur: (e: React.FocusEvent<HTMLInputElement>) => void;
+//   };
+//   form?: FormikProps<Record<string, any>>;
+//   id?: string;
+//   type?: string;
+//   placeholder?: string;
+//   required?: boolean;
+// };
